refactor: findSalesFiles function and update file paths
